requirements.txt
langdetect
# Core
pandas>=2.0.3
numpy>=1.24.3
scikit-learn>=1.3.0
scipy
python-dotenv>=1.0.0
pydantic
tqdm
matplotlib
textstat

# LLM / LangChain (optional backends)
langchain
langchain-community
openai>=1.40.0
anthropic>=0.34.0
mistralai>=0.4.0

# Embeddings / Retrieval
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# PDF / Parsing
pdfminer.six
PyMuPDF
pdfplumber
beautifulsoup4
regex>=2023.5.5

# Web UI (choose one; Streamlit app provided)
streamlit>=1.29.0
gradio

# API (optional)
fastapi
uvicorn

# Export
reportlab
fpdf==2.5.7
gdown>=4.7.1
